requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
pdf2docx==0.5.6
python-docx2pdf==0.1.8
Pillow==8.4.0
Werkzeug==2.0.1
WTForms==2.3.3